Remove the while loop used to run a siddhi app
With this while loop a siddhi app is used to to run continuously. As a
result of that #9 occurred. With the previous implementation siddhi app
used to run smoothly with an input file because the while loop breaks
inside the input file availability checking if condition. Now by
removing it siddhi app runs in with or without an input file.

Resolves: #9